Update JUnit Tests for List, Sort, Search
